requirements.txt
qq-botpy
watchdog~=4.0.2
requests~=2.31.0
pyyaml~=6.0.1
aiohttp~=3.9.5
httpx~=0.27.0
tqdm~=4.66.5
distro~=1.9.0
psutil~=6.0.0
pymongo~=4.8.0
elasticsearch~=8.15.0
ruamel.yaml
jieba~=0.42.1
scikit-learn~=1.5.1
keyboard~=0.13.5